requirements.txt
apprise
python-logging-loki
python-dotenv
click
PyYAML
python-dotenv
